% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigsolve.R
\name{sig_solve}
\alias{sig_solve}
\title{Signature fitting}
\usage{
sig_solve(catalogue, signatures, method = "nnls")
}
\arguments{
\item{catalogue}{catalogue of mutations observed in your sample. A named vector where names are channel and values are counts.}

\item{signatures}{a channel X signature matrix where values represent fractional contributions of each channels.}

\item{method}{how should fitting be performed.
\strong{nnls}: Use non-negative least squares method}
}
\value{
a named vector with 1 element per signature. Values indicate the estimated number of mutation drawn from that signature.
}
\description{
Identify the most likely combination of known mutational signatures
which explain the mutational profile observed in a sample of interest.
}
\examples{
observed_counts <- c(
"C>A" = 183,
"C>G" = 779,
"C>T" = 588,
"T>A" = 706,
"T>C" = 384,
"T>G" = 127
)

signatures <- simulate_signature_matrix(
  signatures = c("sig1", "sig2", "sig3", "sig4"),
  channels = c("C>A", "C>G", "C>T", "T>A", "T>C", "T>G"),
  seed=1
)


# fitting using nnls method
sig_solve(observed_counts, signatures, method = "nnls")

}
